Added [COMPLETE] Problem 2
The goal of this problem was to determine the sum of all the terms in the Fibonacci sequence that are even and under 4 million in value. This program has a function that recursively calculates the terms in the Fibonacci sequence. Afterwards, there is a combination of a for loop and if statements to calculate those values that are even and under 4 million, which are then added to the sum. Lastly, the sum is printed to the console.